+++ src/mdformat/_cli.py
@@ -1,30 +1,26 @@
 from __future__ import annotations
 
 import argparse
-from collections.abc import Callable, Generator, Iterable, Mapping, Sequence
+from collections.abc import Generator, Iterable, Mapping, Sequence
 import contextlib
-import itertools
+import inspect
 import logging
+import os.path
 from pathlib import Path
-import re
 import shutil
 import sys
 import textwrap
 
 import mdformat
-from mdformat._compat import importlib_metadata
 from mdformat._conf import DEFAULT_OPTS, InvalidConfError, read_toml_opts
-from mdformat._util import atomic_write, detect_newline_type, is_md_equal
+from mdformat._util import detect_newline_type, is_md_equal
 import mdformat.plugins
 import mdformat.renderer
-
-# Match "\r" and "\n" characters that are not part of a "\r\n" sequence
-RE_NON_CRLF_LINE_END = re.compile(r"(?:[^\r]|^)\n|\r(?:[^\n]|\Z)")
 
 
 class RendererWarningPrinter(logging.Handler):
     def emit(self, record: logging.LogRecord) -> None:
-        if record.levelno >= logging.WARNING:
+        if record.levelno >= logging.WARNING:  # pragma: no branch
             sys.stderr.write(f"Warning: {record.msg}\n")
 
 
@@ -39,10 +35,16 @@ def run(cli_args: Sequence[str]) -> int:  # noqa: C901
         for plugin in enabled_parserplugins.values()
     )
 
-    arg_parser = make_arg_parser(enabled_parserplugins, enabled_codeformatters)
+    arg_parser = make_arg_parser(
+        mdformat.plugins._PARSER_EXTENSION_DISTS,
+        mdformat.plugins._CODEFORMATTER_DISTS,
+        enabled_parserplugins,
+    )
     cli_opts = {
         k: v for k, v in vars(arg_parser.parse_args(cli_args)).items() if v is not None
     }
+    cli_core_opts, cli_plugin_opts = separate_core_and_plugin_opts(cli_opts)
+
     if not cli_opts["paths"]:
         print_paragraphs(["No files have been passed in. Doing nothing."])
         return 0
@@ -56,11 +58,31 @@ def run(cli_args: Sequence[str]) -> int:  # noqa: C901
     renderer_warning_printer = RendererWarningPrinter()
     for path in file_paths:
         try:
-            toml_opts = read_toml_opts(path.parent if path else Path.cwd())
+            toml_opts, toml_path = read_toml_opts(path.parent if path else Path.cwd())
         except InvalidConfError as e:
             print_error(str(e))
             return 1
-        opts: Mapping = {**DEFAULT_OPTS, **toml_opts, **cli_opts}
+
+        opts: Mapping = {**DEFAULT_OPTS, **toml_opts, **cli_core_opts}
+        for plugin_id, plugin_opts in cli_plugin_opts.items():
+            if plugin_id in opts["plugin"]:
+                opts["plugin"][plugin_id] |= plugin_opts
+            else:
+                opts["plugin"][plugin_id] = plugin_opts
+
+        if sys.version_info >= (3, 13):  # pragma: >=3.13 cover
+            if is_excluded(path, opts["exclude"], toml_path, "exclude" in cli_opts):
+                continue
+        else:  # pragma: <3.13 cover
+            if "exclude" in toml_opts:
+                print_error(
+                    "'exclude' patterns are only available on Python 3.13+.",
+                    paragraphs=[
+                        "Please remove the 'exclude' list from your .mdformat.toml"
+                        " or upgrade Python version."
+                    ],
+                )
+                return 1
 
         if path:
             path_str = str(path)
@@ -82,14 +104,10 @@ def run(cli_args: Sequence[str]) -> int:  # noqa: C901
             _filename=path_str,
         )
         newline = detect_newline_type(original_str, opts["end_of_line"])
+        formatted_str = formatted_str.replace("\n", newline)
 
         if opts["check"]:
-            original_str_lf_eol = original_str.replace("\r\n", "\n").replace("\r", "\n")
-            if (
-                (formatted_str != original_str_lf_eol)
-                or (newline == "\n" and "\r" in original_str)
-                or (newline == "\r\n" and RE_NON_CRLF_LINE_END.search(original_str))
-            ):
+            if formatted_str != original_str:
                 format_errors_found = True
                 print_error(f'File "{path_str}" is not formatted.')
         else:
@@ -103,20 +121,22 @@ def run(cli_args: Sequence[str]) -> int:  # noqa: C901
                 print_error(
                     f'Could not format "{path_str}".',
                     paragraphs=[
-                        "The formatted Markdown renders to different HTML than the input Markdown. "  # noqa: E501
-                        "This is likely a bug in mdformat. "
-                        "Please create an issue report here, including the input Markdown: "  # noqa: E501
-                        "https://github.com/executablebooks/mdformat/issues",
+                        "Formatted Markdown renders to different HTML than input Markdown. "  # noqa: E501
+                        "This is a bug in mdformat or one of its installed plugins. "
+                        "Please retry without any plugins installed. "
+                        "If this error persists, "
+                        "report an issue including the input Markdown "
+                        "on https://github.com/executablebooks/mdformat/issues. "
+                        "If not, "
+                        "report an issue on the malfunctioning plugin's issue tracker.",
                     ],
                 )
                 return 1
             if path:
-                atomic_write(path, formatted_str, newline)
+                if formatted_str != original_str:
+                    path.write_bytes(formatted_str.encode())
             else:
-                with open(
-                    sys.stdout.fileno(), "w", closefd=False, newline=newline
-                ) as stdout:
-                    stdout.write(formatted_str)
+                sys.stdout.buffer.write(formatted_str.encode())
     if format_errors_found:
         return 1
     return 0
@@ -132,23 +152,26 @@ def validate_wrap_arg(value: str) -> str | int:
 
 
 def make_arg_parser(
+    parser_extension_dists: Mapping[str, tuple[str, list[str]]],
+    codeformatter_dists: Mapping[str, tuple[str, list[str]]],
     parser_extensions: Mapping[str, mdformat.plugins.ParserExtensionInterface],
-    codeformatters: Mapping[str, Callable[[str, str], str]],
 ) -> argparse.ArgumentParser:
-    plugin_versions_str = get_plugin_versions_str(parser_extensions, codeformatters)
+    epilog = get_plugin_info_str(parser_extension_dists, codeformatter_dists)
     parser = argparse.ArgumentParser(
         description="CommonMark compliant Markdown formatter",
-        epilog=f"Installed plugins: {plugin_versions_str}"
-        if plugin_versions_str
-        else None,
+        epilog=(epilog if epilog else None),
+        formatter_class=argparse.RawDescriptionHelpFormatter,
     )
     parser.add_argument("paths", nargs="*", help="files to format")
     parser.add_argument(
         "--check", action="store_true", help="do not apply changes to files"
     )
     version_str = f"mdformat {mdformat.__version__}"
-    if plugin_versions_str:
-        version_str += f" ({plugin_versions_str})"
+    plugin_version_str = get_plugin_version_str(
+        {**parser_extension_dists, **codeformatter_dists}
+    )
+    if plugin_version_str:
+        version_str += f" ({plugin_version_str})"
     parser.add_argument("--version", action="version", version=version_str)
     parser.add_argument(
         "--number",
@@ -167,10 +190,56 @@ def make_arg_parser(
         choices=("lf", "crlf", "keep"),
         help="output file line ending mode (default: lf)",
     )
+    if sys.version_info >= (3, 13):  # pragma: >=3.13 cover
+        parser.add_argument(
+            "--exclude",
+            action="append",
+            metavar="PATTERN",
+            help="exclude files that match the Unix-style glob pattern "
+            "(multiple allowed)",
+        )
     for plugin in parser_extensions.values():
         if hasattr(plugin, "add_cli_options"):
+            import warnings
+
+            plugin_file, plugin_line = get_source_file_and_line(plugin)
+            warnings.warn_explicit(
+                "`mdformat.plugins.ParserExtensionInterface.add_cli_options`"
+                " is deprecated."
+                " Please use `add_cli_argument_group`.",
+                DeprecationWarning,
+                filename=plugin_file,
+                lineno=plugin_line,
+            )
             plugin.add_cli_options(parser)
+    for plugin_id, plugin in parser_extensions.items():
+        if hasattr(plugin, "add_cli_argument_group"):
+            group = parser.add_argument_group(title=f"{plugin_id} plugin")
+            plugin.add_cli_argument_group(group)
+            for action in group._group_actions:
+                action.dest = f"plugin.{plugin_id}.{action.dest}"
     return parser
+
+
+def separate_core_and_plugin_opts(opts: Mapping) -> tuple[dict, dict]:
+    """Move dotted keys like 'plugin.gfm.some_key' to a separate mapping.
+
+    Return a tuple of two mappings. First is for core CLI options, the
+    second for plugin options. E.g. 'plugin.gfm.some_key' belongs to the
+    second mapping under {"gfm": {"some_key": <value>}}.
+    """
+    cli_core_opts = {}
+    cli_plugin_opts: dict = {}
+    for k, v in opts.items():
+        if k.startswith("plugin."):
+            _, plugin_id, plugin_conf_key = k.split(".", maxsplit=2)
+            if plugin_id in cli_plugin_opts:
+                cli_plugin_opts[plugin_id][plugin_conf_key] = v
+            else:
+                cli_plugin_opts[plugin_id] = {plugin_conf_key: v}
+        else:
+            cli_core_opts[k] = v
+    return cli_core_opts, cli_plugin_opts
 
 
 class InvalidPath(Exception):
@@ -183,10 +252,9 @@ class InvalidPath(Exception):
 def resolve_file_paths(path_strings: Iterable[str]) -> list[None | Path]:
     """Resolve pathlib.Path objects from filepath strings.
 
-    Convert path strings to pathlib.Path objects. Resolve symlinks.
-    Check that all paths are either files, directories or stdin. If not,
-    raise InvalidPath. Resolve directory paths to a list of file paths
-    (ending with ".md").
+    Convert path strings to pathlib.Path objects. Check that all paths
+    are either files, directories or stdin. If not, raise InvalidPath.
+    Resolve directory paths to a list of file paths (ending with ".md").
     """
     file_paths: list[None | Path] = []  # Path to file or None for stdin/stdout
     for path_str in path_strings:
@@ -194,23 +262,53 @@ def resolve_file_paths(path_strings: Iterable[str]) -> list[None | Path]:
             file_paths.append(None)
             continue
         path_obj = Path(path_str)
-        path_obj = _resolve_path(path_obj)
+        path_obj = _normalize_path(path_obj)
         if path_obj.is_dir():
             for p in path_obj.glob("**/*.md"):
-                p = _resolve_path(p)
-                file_paths.append(p)
-        else:
+                if p.is_file():
+                    p = _normalize_path(p)
+                    file_paths.append(p)
+        elif path_obj.is_file():  # pragma: nt no cover
             file_paths.append(path_obj)
+        else:  # pragma: nt no cover
+            raise InvalidPath(path_obj)
     return file_paths
 
 
-def _resolve_path(path: Path) -> Path:
-    """Resolve path.
+def is_excluded(  # pragma: >=3.13 cover
+    path: Path | None,
+    patterns: list[str],
+    toml_path: Path | None,
+    excludes_from_cli: bool,
+) -> bool:
+    if not path:
+        return False
 
-    Resolve symlinks. Raise `InvalidPath` if the path does not exist.
-    """
+    if not excludes_from_cli and toml_path:
+        exclude_root = toml_path.parent
+    else:
+        exclude_root = Path.cwd()
+
     try:
-        path = path.resolve()  # resolve symlinks
+        relative_path = path.relative_to(exclude_root)
+    except ValueError:
+        return False
+
+    return any(
+        relative_path.full_match(pattern)  # type: ignore[attr-defined]
+        for pattern in patterns
+    )
+
+
+def _normalize_path(path: Path) -> Path:
+    """Normalize path.
+
+    Make the path absolute, resolve any ".." sequences. Do not resolve
+    symlinks, as it would interfere with 'exclude' patterns. Raise
+    `InvalidPath` if the path does not exist.
+    """
+    path = Path(os.path.abspath(path))
+    try:
         path_exists = path.exists()
     except OSError:  # Catch "OSError: [WinError 123]" on Windows  # pragma: no cover
         path_exists = False
@@ -261,35 +359,49 @@ def log_handler_applied(
         logger.removeHandler(handler)
 
 
-def get_package_name(obj: object) -> str:
-    # Packages and modules should have `__package__`
-    if hasattr(obj, "__package__"):
-        package_name = obj.__package__  # type: ignore[attr-defined]
-    else:  # class or function
-        module_name = obj.__module__
-        package_name = module_name.split(".", maxsplit=1)[0]
-    return package_name
+def get_package_name(obj: object) -> str | None:
+    """Return top level module name, or None if not found."""
+    module = inspect.getmodule(obj)
+    return module.__name__.split(".", maxsplit=1)[0] if module else None
 
 
-def get_plugin_versions(
-    parser_extensions: Mapping[str, mdformat.plugins.ParserExtensionInterface],
-    codeformatters: Mapping[str, Callable[[str, str], str]],
-) -> dict[str, str]:
-    versions = {}
-    for iface in itertools.chain(parser_extensions.values(), codeformatters.values()):
-        package_name = get_package_name(iface)
-        try:
-            package_version = importlib_metadata.version(package_name)
-        except importlib_metadata.PackageNotFoundError:
-            # In test scenarios the package may not exist
-            package_version = "unknown"
-        versions[package_name] = package_version
-    return versions
-
-
-def get_plugin_versions_str(
-    parser_extensions: Mapping[str, mdformat.plugins.ParserExtensionInterface],
-    codeformatters: Mapping[str, Callable[[str, str], str]],
+def get_plugin_info_str(
+    parser_extension_dists: Mapping[str, tuple[str, list[str]]],
+    codeformatter_dists: Mapping[str, tuple[str, list[str]]],
 ) -> str:
-    plugin_versions = get_plugin_versions(parser_extensions, codeformatters)
-    return ", ".join(f"{name}: {version}" for name, version in plugin_versions.items())
+    info = ""
+    if codeformatter_dists:
+        info += "installed codeformatters:"
+        for dist, dist_info in codeformatter_dists.items():
+            langs = ", ".join(dist_info[1])
+            info += f"\n  {dist}: {langs}"
+    if parser_extension_dists:
+        if info:
+            info += "\n\n"
+        info += "installed extensions:"
+        for dist, dist_info in parser_extension_dists.items():
+            extensions = ", ".join(dist_info[1])
+            info += f"\n  {dist}: {extensions}"
+    return info
+
+
+def get_plugin_version_str(dist_map: Mapping[str, tuple[str, list[str]]]) -> str:
+    return ", ".join(
+        f"{dist_name} {dist_info[0]}" for dist_name, dist_info in dist_map.items()
+    )
+
+
+def get_source_file_and_line(obj: object) -> tuple[str, int]:
+    import inspect
+
+    try:
+        filename = inspect.getsourcefile(obj)  # type: ignore[arg-type]
+        if filename is None:  # pragma: no cover
+            filename = "not found"
+    except TypeError:  # pragma: no cover
+        filename = "built-in object"
+    try:
+        _, lineno = inspect.getsourcelines(obj)  # type: ignore[arg-type]
+    except (OSError, TypeError):  # pragma: no cover
+        lineno = 0
+    return filename, lineno