+++ tests/test_config_file.py
@@ -6,11 +6,12 @@ import pytest
 
 from mdformat._cli import run
 from mdformat._conf import read_toml_opts
+from tests.test_cli import FORMATTED_MARKDOWN, UNFORMATTED_MARKDOWN
 
 
 def test_cli_override(tmp_path):
     config_path = tmp_path / ".mdformat.toml"
-    config_path.write_text("wrap = 'no'")
+    config_path.write_text("wrap = 'no'\nend_of_line = 'lf'")
 
     file_path = tmp_path / "test_markdown.md"
     file_path.write_text("remove\nthis\nwrap\n")
@@ -65,9 +66,15 @@ def test_invalid_toml(tmp_path, capsys):
         ("wrap", "wrap = -3"),
         ("end_of_line", "end_of_line = 'lol'"),
         ("number", "number = 0"),
+        ("exclude", "exclude = '**'"),
+        ("exclude", "exclude = ['1',3]"),
+        ("plugin", "plugin = []"),
+        ("plugin", "plugin.gfm = {}\nplugin.myst = 1"),
     ],
 )
 def test_invalid_conf_value(bad_conf, conf_key, tmp_path, capsys):
+    if conf_key == "exclude" and sys.version_info < (3, 13):
+        pytest.skip("exclude conf only on Python 3.13+")
     config_path = tmp_path / ".mdformat.toml"
     config_path.write_text(bad_conf)
 
@@ -91,3 +98,53 @@ def test_conf_with_stdin(tmp_path, capfd, monkeypatch):
         assert run(("-",)) == 0
     captured = capfd.readouterr()
     assert captured.out == "1. one\n2. two\n3. three\n"
+
+
+@pytest.mark.skipif(
+    sys.version_info >= (3, 13), reason="'exclude' only possible on 3.13+"
+)
+def test_exclude_conf_on_old_python(tmp_path, capsys):
+    config_path = tmp_path / ".mdformat.toml"
+    config_path.write_text("exclude = ['**']")
+
+    file_path = tmp_path / "test_markdown.md"
+    file_path.write_text("# Test Markdown")
+
+    assert run((str(file_path),)) == 1
+    assert "only available on Python 3.13+" in capsys.readouterr().err
+
+
+@pytest.mark.skipif(
+    sys.version_info < (3, 13), reason="'exclude' only possible on 3.13+"
+)
+def test_exclude(tmp_path, capsys):
+    config_path = tmp_path / ".mdformat.toml"
+    config_path.write_text("exclude = ['dir1/*', 'file1.md']")
+
+    dir1_path = tmp_path / "dir1"
+    file1_path = tmp_path / "file1.md"
+    file2_path = tmp_path / "file2.md"
+    file3_path = tmp_path / dir1_path / "file3.md"
+    dir1_path.mkdir()
+    file1_path.write_text(UNFORMATTED_MARKDOWN)
+    file2_path.write_text(UNFORMATTED_MARKDOWN)
+    file3_path.write_text(UNFORMATTED_MARKDOWN)
+
+    assert run((str(tmp_path),)) == 0
+    assert file1_path.read_text() == UNFORMATTED_MARKDOWN
+    assert file2_path.read_text() == FORMATTED_MARKDOWN
+    assert file3_path.read_text() == UNFORMATTED_MARKDOWN
+
+
+@pytest.mark.skipif(
+    sys.version_info < (3, 13), reason="'exclude' only possible on 3.13+"
+)
+def test_empty_exclude(tmp_path, capsys):
+    config_path = tmp_path / ".mdformat.toml"
+    config_path.write_text("exclude = []")
+
+    file1_path = tmp_path / "file1.md"
+    file1_path.write_text(UNFORMATTED_MARKDOWN)
+
+    assert run((str(tmp_path),)) == 0
+    assert file1_path.read_text() == FORMATTED_MARKDOWN